fonts: Add 6x10 font
This font is suitable for framebuffer consoles on mobile devices with
a 320x240 screen, to get a reasonable number of characters (53x24)
that are still at a readable size.

The font is derived from the existing 6x11 font, but gets 3 extra lines
without sacrificing readability. Also I redesigned a some glyhps so they
are more distinct and better fill the available space.