fb: Add JZ4770 framebuffer driver from Ingenic
Based on driver from linux-2.6.31.3-jz-20110809-FR1.patch.gz.
Only the drivers relevant to JZ4770 are included.
Stripped functionality to what we really need, to make the code more
suitable for restructuring.

Removed all panel support code, this is handled via separate panel
drivers now.

Implemented the check_var function, now SDL applications start up.
This implementation only supports the hardcoded 320x240@16bpp and
320x240@32bpp video modes.

The mmap-ed frame buffer is configured to be fully cacheable. This
improves the raw pixel writing rate by about 200% compared to an
uncached frame buffer. The cache is flushed when a new page starts
being displayed.

Implemented blanking.

Wait-for-vsync is implemented both as an explicit ioctl and implicitly
when changing the vertical panning. The latter allows a more reliable
page flip in combination with a patched SDL.

The driver now always uses the foreground 1 layer in OSD mode. This is
necessary in preparation for TV-out support. Also it will allow use of
the foreground 0 layer for a splash screen and to show OSD messages.

The LCD timing is configured to get a frame rate of exactly 60 Hz.
The panel timings are a bit unconventional, but it works fine with our
panel and it avoids stutters when applications wait both on timers and
on vsync. These timings might not work with TV-out, but we'll deal with
that once TV-out is actually implemented.

I set myself as the author, since this driver is now very different
from the original driver it was based on, so the original author cannot
be expected to answer questions about it. Proper credit is given in the
comment at the top of the file.