fb: Add JZ4770 framebuffer driver from Ingenic
Based on driver from linux-2.6.31.3-jz-20110809-FR1.patch.gz.
Only the drivers relevant to JZ4770 are included.
Stripped functionality to what we really need, to make the code more
suitable for restructuring.

Removed all panel support code, this is handled via separate panel
drivers now.

Implemented the check_var function, now SDL applications start up.
This implementation only supports the hardcoded 320x240@32bpp video
mode.

Implemented blanking.

The driver now always uses the foreground 1 layer in OSD mode. This is
necessary in preparation for TV-out support. Also it will allow use of
the foreground 0 layer for a splash screen and to show OSD messages.

I set myself as the author, since this driver is now very different
from the original driver it was based on, so the original author cannot
be expected to answer questions about it. Proper credit is given in the
comment at the top of the file.