i2c: Add JZ4770 I2C driver from Ingenic
Taken from linux-2.6.31.3-jz-20110809-FR1.patch.gz.
Only the drivers relevant to JZ4770 are included.

Use one driver with 3 devices instead of 3 drivers, to better fit
the Linux driver model. Also do not pretend that i2c2 is interface 5.

The board decides whether to use DMA or not, instead of the kernel
config.