MIPS: JZ4770: GCW0: Added defconfig
This is similar to the OpenDingux Dingoo A320 defconfig, but adapted
to the GCW Zero hardware.

There used to be a qemu config as well, but that was dropped in
jz-3.17 because it was unmaintained. It was still present in jz-3.16
but not fully working anymore at that point.

In 3.17 the following options were disabled under the assumption no
applications depend on them: CONFIG_SYSVIPC and CONFIG_LEGACY_PTYS.

Squashed in transition from 3.19 to 4.0:
- USB gadget started using configfs
- put mininit on system partition instead of in initframfs
- enabled kexec syscall