ASoC: Add JZ4770 ASoC support
The hardware-facing code comes from linux-2.6.31.3-jz-20110809-FR1.patch.gz
as published by Ingenic. The Linux-facing code was mostly rewritten by me.

Period and buffer sizes are set to the same values as on JZ4740.
The original sizes lead to an audio latency of almost a second, which
is unacceptably high for gaming.