ASoC: JZ4770: Add GCW0 board driver
This adds support for the GCW Zero handheld game console.

The "Headphone" widget name matches the name in the codec, such that
alsamixer will display a combined widget for the headphones volume and
mute.

The 50 ms delay on HP amp power up was determined experimentally:
I couldn't hear a difference when increasing the the delay, while the
pop when using a delay of 10 ms was slightly louder.