fb: JZ4770: Switch to using foreground 1 layer in OSD mode
Previously, non-OSD mode was used. However, in the future we want to
be able to show OSD messages from pwswd or perhaps a virtual keyboard.
Also the OSD is a good way to show a splash screen without the logo
trickery we previously used.

Additionally, the IPU can output to foreground layer 1, which would
enable us to use hardware scaling.

Finally, TV-out seems to work only in OSD mode: in non-OSD mode the
colors are wrong no matter how I configure the LCDC. Possibly the TV
encoder only accepts YUV input and our frame buffer is in RGB; there
is a flag to convert RGB to YUV but that flag only works in OSD mode.