video: displays: nt39016: Go into standby mode during blanking
This should save some power when the LCDC and the panel itself are
powered down.

While the diff looks large, there are only two changes in the init
sequence: reg 0 is written (to get out of standby mode) and reg 0x1E
is skipped (it's part of a programming sequence; the previous value
was safe but not writing it at all is even safer).