MIPS: JZ4770: GCW0: Eliminate CONFIG_JZ_SYSTEM_AT_CARD
Since we had that feature enabled in our config, keep the "defined"
version of each check.

Removed commented out code to wake up on SD insertion. Inserting a card
can mean the user wants to access files from that card, but it could
just as well mean the user wants to store the card inside the unit
without wanting to access it at this time.