net: wireless: rtl8192cu: Import driver from Realtek
There is already a driver for the RTL8192CU in the mainline kernel,
but it is unreliable in our tests. This commit adds a different driver
from Realtek itself, with updates from various people.

Source: https://github.com/dz0ny/rt8192cu/

Removed CR end-of-line characters and a few other unprintable
characters from the sources.
Selected CONFIG_PLATFORM_MIPS_JZ4760 in the Makefile.
Appended "_ALT" to the main config variable name "CONFIG_RTL8192CU", so
this driver can exist in the same source tree as the mainline driver.
Fixed usage of kthread_run.
Update for procfs changes in Linux 3.10.
Fixed out-of-bounds array init.