fb: JZ4770: Re-use the first line of the frame as the "blank line"
The "blank line" is an extra line of pixels, transferred after each
frame. Its sole purpose is to create a delay between the End Of Frame
and the Start Of Frame interrupts, so that vertical synchronization
can be implemented (with only one DMA descriptor, the EOF and SOF
interrupts occur at the exact same moment).

Previously, the "blank line" corresponded to an allocated buffer.
This was not required, as the content of that extra line doesn't
matter: it is never shown on the screen. Thus, we now use the first
line of the frame as the "blank line", so that no extra allocation
is needed.