fb: JZ4770: Cleaned up enable/disable code
Always put a delay between enabling the clock and enabling the LCDC.
At boot this was essential to avoid hangups, so it seems unlikely it
is optional at other times.

When blanking, also shut down LCD panel.

On mode changes, wait for the frame to end before shutting down the
LCDC (normal disable instead of quick disable).