fb: JZ4770: Vsync cleanups
Use the EOF interrupt of the frame descriptor rather than the SOF
interrupt of the dummy line. Although they happen at almost the same
time, it's more intuitive to trigger vsync from an EOF interrupt.

Rewrite the address of the next DMA descriptor to load on every LCDC
enable. Without this, it continues where it left off. This usually
puts the dummy line first, but could potentially put the frame first,
which would shift the entire screen down by one line.

Added a comment explaining why the dummy line is important, even now
that we no longer have an interrupt attached to it.