mmc: JZ4770: Renamed platform data header
The header contains only the platform data struct definition at this
point, so renamed it to set it apart from the other headers that still
contain register definitions as well.

When migrating to Linux 3.10, this header should become part of the
JZ4770 MMC driver, while currently it is introduced as part of the
platform.