fb: JZ4770: Added a timeout on wait-for-vsync
This ensures that if vsync doesn't come, for example because the
display is blanked, the wait doesn't take overly long. The return
value will indicate a timeout happened, so the caller knows that
there was no vsync.

I set the timeout to 100 ms, which is significantly more than the next
vsync should take, but still a short wait from a user's perspective.