mmc: JZ4770: Use standard name for DMA descriptor struct
Instead of using a macro, use the name directly, including "struct".
This is required now that the platform header no longer has a typedef.

Also moved inclusion of the DMA header to jz_mmc_dma.c, since that is
the only source file that actually has to know about DMA.