USB: ohci-jz4740: Add support for JZ4770
The JZ4770 has additional control bits in the CPM to disable the UHC
PHY when it's not in use. These are in the disabled state at boot, so
we must enable them before we can use the UHC.

Commits squashed from jz-3.12 branch:

Tentative fix for hub control
Needs verification by larsc.