fb: JZ4770: Fix hang at boot if LCDC clock is initially stopped
Disabling the LCDC when the clock is not running hangs the system.
In our use case the LCDC is already disabled when the driver inits,
so rather than doing more complex trickery I just documented that
assumption and removed the disable code.

Clearing the status register is now done on enable, which is probably
better anyway, since we're not interested in status updates from long
ago.