USB: musb: JZ4770: Cleanup and bugfix
Use the same platform driver approach that most musb glue uses.

Use devm. And even with that, explicitly free the IRQ, since we have to
be sure it is no longer active before we remove the timer.

Don't pass IRQF_DISABLED, it's deprecated and currently a NOP.

The last argument when freeing an IRQ must match the last argument of
the request call, not the name. Because of this, the original code
failed to free the IRQ.