rtc: JZ4770: Don't set OPCR_ERCS via cpm_set_clock()
Instead, do the register write directly in the RTC driver.
I'm not sure it belongs there, but at least it's explicit now.

This also removes the "WARNING: can NOT get clock 4119!" message from
the kernel log at boot, since cpm_set_clock() calls cpm_get_clock() at
the end to retrieve the actual clock and that is not supported for
CGU_RTCCLK.