MIPS: JZ4770: GCW0: Model the WiFi power pin as GPIO regulator for vbus
Since the WiFi module is the only device on the bus, power for the bus
and power for the device are the same thing.

Note that this pin disables the entire WiFi module, not just the radio
part, so rfkill is not the right way of modeling it in the kernel.