ASoC: JZ4770: codec: Force DAC to be enabled during playback
In commit e415baca I wrote that there is no need to keep the DAC active
when all outputs are disabled, but that turns out to be false: if the
DAC is powered down, the codec will no longer consume samples from the
AIC and as a result user space can block on the delivery of samples.

I solved this by making the SYSCLK output explicit and forcing this
output to be enabled during playback, which in turn keeps the DAC
powered.

The capture stream might require a similar construct, but for the
moment I'm only interested in playback.

Thanks to Lars-Peter Clausen for his help in figuring out the problem.