fb: JZ4770: Improved video memory allocation
Added support for multiple video modes. Enough space is reserved for
the largest mode at 32 bpp and with double buffering.

Cleaned up existing code. The unmap code was freeing things that were
never reserved (in this version of the code, anyway), so that would
probably have crashed and burned if it was ever called.