gpu: vivante: Moved debug levels to internal options
There were two places in the uapi where enabling debugging would add
fields to the structs. This is of course undesired since it creates a
different ABI when debugging. However, the extra fields that were added
weren't even written by the driver. So I removed them.

With those fields gone, the debug level didn't affect the uapi anymore,
so I moved the relevant defines to the internal options header.