jz4780-tcu: driver for the Ingenic jz4780 Timer/Counter Unit
This patch provides a driver for the Ingenic jz4780 Timer/Counter Unit
(TCU). An interface is provided with which timer channels can be
requested and configured, and atop that support for using channels as
the clocksource or as clock event devices is implemented.

TODO: split out jz47xx & jz4780 to multiple commits?
TODO: DT binding doc

Signed-off-by: Paul Burton <paul.burton@imgtec.com>