fb: JZ4770: Upload frames using the IPU instead of DMA
The IPU is the Image Processing Unit. It handles scaling with bilinear
or bicubic filtering as well as colorspace conversion.
With this change, the LCD controller does not use the DMA anymore to
upload the framebuffer to the screen; instead, the framebuffer is read
by the IPU, eventually processed, and uploaded directly to the LCD
without crossing the RAM again.

For now, this brand new video driver handles all the features of the
previous one:
- Single buffering with deferred cache flush,
- VSYNC with perfect 60 Hz,
- Double / triple buffering,
- 32 / 16 bpp,
- Blanking, suspend / resume.

Some differences:
- The LCD is always configured for 32 bpp; when a different mode is
  used, the IPU is used to convert the framebuffer to 32 bpp.
- Minor regression: as the IPU takes some time to configure (and will
  take even more when scaling is implemented), garbage is displayed for
  a couple of milliseconds before the first image is displayed.