watchdog: JZ4740: Fix incorrect register write for stopping the watchdog
It was doing a 16-bit access to set the bit 16, so the value was never
written. Plus, only the bit 16 should be set, as setting other bits may
result in having other clocks started or stopped. Finally, it is not
possible to access the registers once the watchdog clock is stopped, so
we now stop it at the last moment.

This fixes commit 30c7a8d6