input: linkdev: Send key release events when changing key map
When switching between the regular and the alternative key maps,
it is required that the input device does not report any of its keys
as pressed, otherwise the input handlers get confused. To avoid that,
we send a "key released" event for each of the keys we support, even
if the key isn't pressed in the first place.