fb: JZ4770: Allow hot-switching the "keep aspect ratio" param at runtime
This change makes sure that the IPU gets reconfigured as soon as the
keep_aspect_ratio parameter is written in sysfs.

Note that this commit has a side effect, the sysfs nodes moved in the
sysfs tree from /sys/module/jz4770-fb/parameters to
/sys/devices/platform/jz-lcd.0

For consistency, the allow_downscaling parameter moved as well to this
new directory.