fb: JZ4770:  Fix an issue where x/32 or 32/x ratios were refused
It was only allowed up to x/31 or 31/x before, but the IPU really
accepts up to 32 coefficients.

Furthermore, a ratio of e.g. 32/63 is totally possible, as the
number of coefficients corresponds to the numerator of the ratio.
Now, the code only verifies that the numerator is inferior or
equal to 32.

Additionally, the code now makes use of <linux/gcd.h> instead of
duplicating the functionality in a custom function.