input: pwm-haptic: Don't set polarity at all
If the polarity should be set to a non-default value, it likely means
the motor will be running until it is set. I don't think a hardware
design would want that to happen, so most likely it would either use
the default polarity or it would have a power regulator that is kept
off until the PWM is configured properly. In either case the polarity
support that was present isn't useful, so I'm removing it.

Platform data is being replaced by Device Tree anyway, so I think it's
best to put only values in there that we need right now.