input: pwm-haptic: Clean up probe and remove function
Include error codes in error logging.

Use "idev" instead of "haptic->input_dev" where possible.

Do not call input_ff_destroy:
Calling it when input_ff_create_memless fails is unnecessary because
the "ff" field hasn't been set in that case.
Calling it when input_register_device fails is unnecessary because
the managed destruction of the input_dev will take care of destroying
the ff_device. I verified this by attaching a "destroy" handler.
Calling it in pwm_haptic_remove is unnecessary because the managed
destruction of input_dev will destroy the ff_device as well.