PWM: JZ4770: Correct polarity again
My earlier conclusion (see cb0eddeb6) that the output while disabled is
the non-initial state was incorrect. For PWM4, the level is the inital
state when disabled, consistent with the programming manual.

For PWM1, the output at boot is independent of the initial state
setting. So there is a difference here between TCU1 and TCU2 modes.