fb: JZ4770: Removed panel hooks
We had a custom interface to power on/off the LCD panel. However, the
existing platform-lcd driver can handle this just fine, so this commit
removes the custom interface.

Since the panel hooks were the only contents of the platform data,
the platform data for this driver is removed completely.