serial: 8250_ingenic: Add MCR register bits for flow control
These bits are not standard 550 uart.
For the serial module in the JZ4780, they need to be enabled
for flow control to work.

Found using old ingenic history to get BT to work on 3.16

Based on original patch by Zubair, changed to keep changes local to
the Ingenic UART driver by hooking in to serial_out and doing the
right thing when modem status interrupts are enabled/disabled.

Signed-off-by: Alex Smith <alex.smith@imgtec.com>