pwm: jz4740: Fix duty cycle
The initial level is also the level of the PWM output when the PWM is
disabled. This means the "halfway" point should move forward when the
duty cycle becomes longer. This is how this driver works upstream,
but it got broken in "pwm: jz4740: Use ioremap to map the registers",
in a change that is unrelated to the register mapping.