fb: JZ4770: Properly lock access for par struct and hw regs
Previously, there was a mutex that was only acquired by jzfb_blank(),
so it wasn't doing anything in practice. Since the interrupt handler
accesses both the par struct and the hw regs and mutexes can't be used
in an interrupt context, I changed the mutex into a spinlock. I also
added locking in all operation callbacks that access hw regs or shared
state in the par struct (struct jzfb).