clk: jz4740-tcu: Require gating before parent and rate changes
This simplifies the code a lot: since we know the clk framework will
call set_parent and set_rate with the clock gated, we can ungate the
clock, write the registers and gate it again unconditionally.

The hardware doesn't allow the divider to be changed on a running
timer/counter, so set_rate gating is unavoidable. Reparenting a
running timer/counter is not explicitly disallowed, but it is
likely to at least disrupt the rate at which the counter ticks
for a short time, so it is probably safer to disallow it too.

This commit seems to fix initial rate setting from DT, probably
because forced ungating for set_rate was missing before.