MIPS: JZ4770: Increase zload address
The zload address and the regular load address were picked in a time
when the kernel was smaller and fewer drivers were included. On the
GCW Zero (JZ4770) we sometimes have non-booting kernels when enabling
for example debugfs, caused by the unpacked kernel overwriting the
compressed data and/or unpack code. To avoid such problems, load the
compressed data at 16MB instead of 6MB.