clocksource: JZ4770: Tighten Kconfig dependencies
This avoids offering the JZ4770 OST driver on JZ4740, which does not
have the OST hardware.

MACH_INGENIC determines the availability of jz4740_tcu_write_tcsr.
While MACH_INGENIC is implied by MACH_JZ4770 || MACH_JZ4780, it is
not implied by COMPILE_TEST, so we have to test for it explicitly.