backlight: nt39016: Make it a platform driver module
Previously, the NT39016 support was called from platform_lcd hooks.
Making it into a full driver gives us more options for future changes,
such as converting it into an SPI protocol driver.

I moved all the GCW0 glue into this driver for now. The GPIOs should
be managed by the SPI master, but it is not clear to me yet where the
regulators belong. In any case, we will be using DT instead of
platform data, so I removed the platform data headers.