backlight: nt39016: Improve reset sequence
The regulators control power for the panel itself, not for the
NT39016. So we don't need to reset it on every power up. We still
reset it on probe though, just in case the hardware is in a weird
state.

The reset pulse we sent was too short to actually reset the IC.
See new code comment for details.

I also changed the first value written to the register init sequence
to reset the IC, to send a reset command after the reset pulse.
Probably unnecessary, but it doesn't hurt either.

Replaced use of udelay by usleep_range: even though we don't wait
for very long, there is no reason to use a blocking delay.