fbdev: jz4770_fb: Add bicubic screen-scaling
The IPU supports both bilinear and bicubic screen-scaling modes. Until
now, we've only used bilinear mode, and had no option to support
nearest-neighbor scaling for arbitrary scale factors.

Bicubic mode is sharper and is now used exlusively. Nearest-neighbor and
bilinear modes are emulated via bicubic mode. There is no compelling
reason to keep support for bilinear mode. However, it'd be easy to add
bilinear support back in in the future: we are still doing the same
bilinear coefficient math we always did, as it's an intermediate step in
computing the coefficients for bicubic mode.

Added two sysfs settings 'sharpness_upscaling', 'sharpness_downscaling'.
A value of 0 gives nearest-neighbor scaling, 1 gives bilinear scaling
(both emulated). Values of 2..32 give bicubic with increasing levels of
sharpness. The default values are both 8, translating internally to a
bicubic sharpness factor of -1.0, giving a decent increase in sharpness
to both upscaling and downscaling.

Signed-off-by: Daniel Silsby <dansilsby@gmail.com>